kafka-console-app/requirements.txt
requests
confluent_kafka
asyncio
python-dotenv
